pyproject.toml
[project]
name = "qrmaker"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow",
    "segno"
]